libs: core: add new set of core files
We are still trying to get the core files working with platformio.

This adds a set of core files, but they still fail to build.  It looks
like there may be a stale set of files on the server, but it is unclear
exactly what's going on.

The PXT_SHIMS_BEGIN symbol doesn't seem to get added, or at least isn't
being linked on the server.

More investigation will need to be done.

Signed-off-by: Sean Cross <xobs@kosagi.com>