editor: hex: correctly handle truncated ihex sections
The ihex format used allows for "truncated" ihex sections that are
shorter than the usual 16-bytes.  The current ltcIhexToBinary() function
does not handle these properly.

This hasn't been an issue so far, because we've happily been filling up
the 16-byte limit.

Pad the resulting hex file so that we can handle these short lines.

Signed-off-by: Sean Cross <sean@xobs.io>