blocksprj: don't append source to output binary
pxt supports appending the program source to the end of the output
binary.  It's LZMA-compressed, so it's not large, but it still adds
overhead.

This enables the import of a program back into Makecode, however since
the Chibi Chip does not support reading code back, this extra data is
simply wasted time.

Turn this feature off by default.  This eliminates several seconds off
the load time.

Signed-off-by: Sean Cross <sean@xobs.io>