pxtarget: set flashalign to 16
The "flashalign" feature allows for placing bytecode on an eraseable
boundary.  This may become important in the future, if we support live
code reloading.  However, for now, it simply adds up to 1024 extra bytes
to the code load.

Lower the flashalign down to 16, which is enough to keep the ihex parser
happy.  Any lower, and the ihex parser won't be able to locate the magic
values.

Signed-off-by: Sean Cross <sean@xobs.io>