Slimmer core (#31)
* libs: core: format some C++ source files

In preparation for cleaning up the C++ source, pre-format some of the
files.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: check in generated string files

These appear to be generated, and the contents have changed.  Check in
the updated contents.

Though, perhaps they should be omitted entirely?

Signed-off-by: Sean Cross <sean@xobs.io>

* variant: run source file through formatter

Signed-off-by: Sean Cross <sean@xobs.io>

* editor: hex: correctly handle truncated ihex sections

The ihex format used allows for "truncated" ihex sections that are
shorter than the usual 16-bytes.  The current ltcIhexToBinary() function
does not handle these properly.

This hasn't been an issue so far, because we've happily been filling up
the 16-byte limit.

Pad the resulting hex file so that we can handle these short lines.

Signed-off-by: Sean Cross <sean@xobs.io>

* wip: slimming down

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: move asserts into pxt.h and make them optional

Move a bunch of assert code into pxt.h, and allow us to #ifdef it out.
This will make it easier to shrink the base image.

Signed-off-by: Sean Cross <sean@xobs.io>

* blocksprj: don't append source to output binary

pxt supports appending the program source to the end of the output
binary.  It's LZMA-compressed, so it's not large, but it still adds
overhead.

This enables the import of a program back into Makecode, however since
the Chibi Chip does not support reading code back, this extra data is
simply wasted time.

Turn this feature off by default.  This eliminates several seconds off
the load time.

Signed-off-by: Sean Cross <sean@xobs.io>

* pxtarget: set flashalign to 16

The "flashalign" feature allows for placing bytecode on an eraseable
boundary.  This may become important in the future, if we support live
code reloading.  However, for now, it simply adds up to 1024 extra bytes
to the code load.

Lower the flashalign down to 16, which is enough to keep the ihex parser
happy.  Any lower, and the ihex parser won't be able to locate the magic
values.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove utvector

There is now no more code that relies on these libraries, so remove
them.

Any code that does rely on these functions has been marked as
"unimplemented".

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: unimplement some more refcollection functions

In an effort to remove utvector, stub out the last functions that use
it.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove String class

It is unclear whether the String class will be used at all.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove Array class

The Array class may be useful when doing Javascript programming, but it
appears to not be used in Blocks mode.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: helpers: remove some comented-out code

It's an old panic() implementation.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: pxt: make hasVTable not-inline

The performance increase isn't worth the space tradeoff.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: shrink "Unimplemented" to "Unimpl"

This string will probably never be seen anyway...

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: mark RefRecord_destroy as unimplemented

This doesn't seem to be used, so remove its implementation.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: deduplicate getNumGlobals

There were two copies of this function.  Make only one.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: pxt: fix ut-related compile errors

These errors snuck in when utvector.h was removed.  When doing a clean
rebuild, these files are removed, but when doing an incremental build
they stick around.  So the error wasn't caught.

Completely remove all vector code.

Signed-off-by: Sean Cross <sean@xobs.io>

* Remove text blocks.