Ltc object support (#70)
* libs: core: remove hash/memleak functions

These are useful for debugging, but as of right now they are unused and
can be removed.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove duplicate incr/decr from pxt declaration

The incr/decr functions were declared twice.  Not an error, just
needless duplication of statements.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: make allocate() globally available

Replace the version of allocate() inside variant.cpp with a core version
that's available to everyone.  Define this as a function with "C"
linkage, so that it is available to C functions in the future.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: add string support

This brings back basic "string" support.  This adds 300 bytes to the
load.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove debugmemleak/templatehash/programhash

These functions appear to be unused.

Removing them saves about 40 bytes.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: RefCounted: wrap sanity check in assert()

Asserts are normally disabled, which can save some space when building a
production release.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: pxt: re-enable RefCollection length()/in_range()

To start bringing back array (and object) functionality, re-enable these
functions.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: randomRange: minor formatting cleanup

These are unnecessary, and by removing them we will save about 50 bytes.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: remove debugmemleaks/(template|program)hash

These functions are unused, and removing them saves about 50 bytes.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: break pxt.(h|cpp) into separate files

As part of a commit to "bring back object support", start by separating
the pxt code into various independent source files.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: add auto-generated files

These get updated by "pxt build".

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: RefCollection: save some space for insertAt

The insertAt() function is simply a specialized version of push().  So
call push(), and then re-order the resulting list.

This is not as time-efficient as implementing our own insertAt(), but it
means we don't need to duplicate insertion code.

Signed-off-by: Sean Cross <sean@xobs.io>

* core: remove BBC copyright notice from headers

Instead, rely on LICENSE.txt.

Signed-off-by: Sean Cross <sean@xobs.io>

* core: move RefObject functions to their own file

These functions somehow got stuck in RefRecord.cpp.  Since they are part
of the RefObject object, move them into their own file.

Signed-off-by: Sean Cross <sean@xobs.io>

* core: refmap: remove commented-out code

This code is not likely to be used or useful, so remove it.

Signed-off-by: Sean Cross <sean@xobs.io>

* core: refrecord: remove print function contents

Removing the print function will save us a bit of space, since we don't
always have text output available.

Signed-off-by: Sean Cross <sean@xobs.io>

* core: enums: commit new "const" values

The pxt compiler now defines some values as "const".  Commit these new
values to the repo.

Signed-off-by: Sean Cross <sean@xobs.io>