Trying a higher resolution version of the render.
The 140 px render looks awful. Rendered two versions, one at
270px, the other at 480 px.

According to the email thread 270px is the "native" resolution
but upon inspecting the other existing gifs they are at 480px.

It's a trade-off between frame rate, responsivity, and asset size.

Any comment/suggestion on maximum asset size for the GIF
would be helpful. I was able to tweak the frame rate down on the
480px gif but some of the other loops I'm looking at will need
a higher frame rate which will cause the render to be ~3-4MiB.