Fix twinkle effect (#245)
* libs: core: remove extra "printf" from events spawner

This "printf" was used for debugging, and might cause confusion when
using input events.  It is unnecessary.

Signed-off-by: Sean Cross <sean@xobs.io>

* libs: core: increase stack size for poll stack

The poll stack size was 256 bytes, which was not enough for certain
effects such as "Twinkle".  Increase the stack size to 400 bytes, which
matches the stack size for normal `forever()` threads.

This fixes issue #234.

Signed-off-by: Sean Cross <sean@xobs.io>